+++ sdk/storage/storage-file-share/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 12.2.0 (unreleased)
 
-- Added `deleteIfExists()` to `ShareClient`, `ShareDirectoryClient`, and `ShareFileClient`.
+- Added `exists` and `deleteIfExists()` to `ShareClient`, `ShareDirectoryClient`, and `ShareFileClient`.
 - Added `createIfNotExists()` to `ShareClient` and `ShareDirectoryClient`.
 
 ## 12.1.2 (2020.05)